Patched mocking.mock() to handle standard library functions.
Although standard library functions have the same type as built-in
functions, they cannot be mocked using the same setattr approach.  To fix
this, we patched mocking.mock() to first look for built-in functions, then
to handle non-built-in functions with the same type as built-in functions.
This will catch standard library functions and mock them appropriately.