Reworked mock() to remove duplicated code.
First, is_mocked boolean is used to run the code that saves the state of the
function being mocked.  This code is located in the 'if not' block after the main
if/elif/else block.  Second, the 'true' builtins will still be handled by
'setattr' then a return statement ends the execution of this method.  All other
cases are allowed to fall through to the code that handles the mock_wrapper.