Changed mocking.mock() to accomodate modules that link to other modules.
We ran into a problem mocking os.readlink as in certain unix environments
os links to an outside module, posix.  This resulted in the __dict__ of
readlink remaining unchanged, and thus not mocking the realink function
(which in our environment resides in the posix module).  To patch this,
we added a third argument to mock() that can be used to explicitly
specify the target_module.  A default value of None is provided so all
prior calls on this method are compatible.